MCP3204 and MCP3208
