+++ mrbgems/picoruby-mcp3208/mrbgem.rake
@@ -1,7 +1,7 @@
-MRuby::Gem::Specification.new('picoruby-mcp3204') do |spec|
+MRuby::Gem::Specification.new('picoruby-mcp3208') do |spec|
   spec.license = 'MIT'
   spec.author  = 'HASUMI Hitoshi'
-  spec.summary = 'MCP3204 ADC module'
+  spec.summary = 'MCP3204/3208 ADC module'
 
   spec.add_dependency 'picoruby-machine'
   spec.add_dependency 'picoruby-spi'
